requirements.txt
requests>=2.31
pymupdf>=1.24
pymupdf4llm>=0.0.10
pyyaml>=6.0
